优化目录结构

+++ sample/src/main/java/la/xiong/androidquick/demo/bean/MenuBean.java
@@ -7,14 +7,14 @@ import java.io.Serializable;
 public class MenuBean implements Serializable {
 
 	public TreeNode treeNode;
-	public int currentId;
+	public long currentId;
 	public String name;
-	public int upperId;
+	public long upperId;
 
 	public MenuBean() {
 
 	}
-	public MenuBean(int currentId, String name, int upperId) {
+	public MenuBean(long currentId, String name, long upperId) {
 		this.currentId = currentId;
 		this.name = name;
 		this.upperId = upperId;
